Finish registering Kodi services in __init__.py
